+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/NewEpoch.hs
@@ -17,7 +17,6 @@
 
 module Cardano.Ledger.Conway.Rules.NewEpoch (
   ConwayNEWEPOCH,
-  ConwayNewEpochPredFailure (..),
   ConwayNewEpochEvent (..),
 ) where
 
@@ -54,27 +53,14 @@ import Cardano.Ledger.Slot (EpochNo (EpochNo))
 import Cardano.Ledger.State
 import qualified Cardano.Ledger.Val as Val
 import Control.DeepSeq (NFData)
+import Control.Exception (assert)
 import Control.State.Transition
 import Data.Default (Default (..))
 import qualified Data.Map.Strict as Map
 import Data.Set (Set)
+import Data.Void (Void)
 import GHC.Generics (Generic)
 import Lens.Micro ((%~), (&), (^.))
-
-newtype ConwayNewEpochPredFailure era
-  = CorruptRewardUpdate
-      RewardUpdate -- The reward update which violates an invariant
-  deriving (Generic)
-
-deriving instance Eq (ConwayNewEpochPredFailure era)
-
-deriving instance
-  ( Show (PredicateFailure (EraRule "EPOCH" era))
-  , Show (PredicateFailure (EraRule "RATIFY" era))
-  ) =>
-  Show (ConwayNewEpochPredFailure era)
-
-instance NFData (ConwayNewEpochPredFailure era)
 
 data ConwayNewEpochEvent era
   = DeltaRewardEvent !(Event (EraRule "RUPD" era))
@@ -121,6 +107,8 @@ instance
   , GovState era ~ ConwayGovState era
   , Eq (PredicateFailure (EraRule "RATIFY" era))
   , Show (PredicateFailure (EraRule "RATIFY" era))
+  , Eq (PredicateFailure (ConwayNEWEPOCH era))
+  , Show (PredicateFailure (ConwayNEWEPOCH era))
   ) =>
   STS (ConwayNEWEPOCH era)
   where
@@ -128,7 +116,7 @@ instance
   type Signal (ConwayNEWEPOCH era) = EpochNo
   type Environment (ConwayNEWEPOCH era) = ()
   type BaseM (ConwayNEWEPOCH era) = ShelleyBase
-  type PredicateFailure (ConwayNEWEPOCH era) = ConwayNewEpochPredFailure era
+  type PredicateFailure (ConwayNEWEPOCH era) = Void
   type Event (ConwayNEWEPOCH era) = ConwayNewEpochEvent era
 
   initialRules =
@@ -162,6 +150,8 @@ newEpochTransition ::
   , GovState era ~ ConwayGovState era
   , Eq (PredicateFailure (EraRule "RATIFY" era))
   , Show (PredicateFailure (EraRule "RATIFY" era))
+  , Eq (PredicateFailure (ConwayNEWEPOCH era))
+  , Show (PredicateFailure (ConwayNEWEPOCH era))
   ) =>
   TransitionRule (ConwayNEWEPOCH era)
 newEpochTransition = do
@@ -216,7 +206,7 @@ updateRewards ::
   Rule (ConwayNEWEPOCH era) 'Transition (EpochState era)
 updateRewards es e ru'@(RewardUpdate dt dr rs_ df _) = do
   let totRs = sumRewards (es ^. prevPParamsEpochStateL . ppProtocolVersionL) rs_
-  Val.isZero (dt <> dr <> toDeltaCoin totRs <> df) ?! CorruptRewardUpdate ru'
+   in assert (Val.isZero (dt <> dr <> toDeltaCoin totRs <> df)) (pure ())
   let !(!es', filtered) = applyRUpdFiltered ru' es
   tellEvent $ RestrainedRewards e (frShelleyIgnored filtered) (frUnregistered filtered)
   -- This event (which is only generated once per epoch) must be generated even if the
@@ -226,8 +216,8 @@ updateRewards es e ru'@(RewardUpdate dt dr rs_ df _) = do
 
 instance
   ( STS (ConwayNEWEPOCH era)
-  , PredicateFailure (EraRule "NEWEPOCH" era) ~ ConwayNewEpochPredFailure era
   , Event (EraRule "NEWEPOCH" era) ~ ConwayNewEpochEvent era
+  , PredicateFailure (EraRule "NEWEPOCH" era) ~ PredicateFailure (ConwayNEWEPOCH era)
   ) =>
   Embed (ConwayNEWEPOCH era) (ShelleyTICK era)
   where